*** empty log message ***

+++ plugins/Calendar.php
@@ -6,8 +6,10 @@ class Calendar extends PluginBase
 		if ($this->dbrelease == $this->GetRelease()) return;
 		switch($this->dbrelease)
 		{
-			case(2):
+			case(1):
 				Zymurgy::$db->query("alter table calendar add location varchar(80)") or die("Coulnd't upgrade to Calendar version 2: ".Zymurgy::$db->error());
+			case(2):
+				Zymurgy::$db->query("alter table calendar add instance integer") or die("Coulnd't upgrade to Calendar version 2: ".Zymurgy::$db->error());
 		}
 		$this->CompleteUpgrade();
 	}
@@ -19,7 +21,7 @@ class Calendar extends PluginBase
 
 	function GetRelease()
 	{
-		return 2;
+		return 3;
 	}
 	
 	function GetUninstallSQL()
@@ -148,28 +150,34 @@ class Calendar extends PluginBase
 	
 	function Render()
 	{
+		if($this->GetConfigValue('Date Format') == null)
+		{
+			echo("Calendar plugin instance must be configured before it can be displayed.");
+			die("Calendar plugin instance must be configured before it can be displayed.");
+		}
+		
 		//Begin by deleting items which are over.
 		//Zymurgy::$db->query("delete from calendar where end<(unix_timestamp()+24*3600)");
 		Zymurgy::$db->query("delete from calendar where end<(unix_timestamp()-24*3600)");
 		$cal = array();
 		$iid = 0 + $this->iid;
-		$ri = Zymurgy::$db->query("select * from calendar where (plugin={$this->pid}) and (instance=$iid) order by start");
+		$ri = Zymurgy::$db->query("select * from calendar where (instance=$iid) order by start");
 		while (($row = Zymurgy::$db->fetch_array($ri))!==false)
 		{
 			$out = array();
-			$start = date($this->config['Date Format'],$row['start']);
-			$end = date($this->config['Date Format'],$row['end']);
+			$start = date($this->GetConfigValue('Date Format'),$row['start']);
+			$end = date($this->GetConfigValue('Date Format'),$row['end']);
 			$date = $start;
-			if (($this->config['Allow Multi-day Events'] == 'yes') && ($end != $start))
+			if (($this->GetConfigValue('Allow Multi-day Events') == 'yes') && ($end != $start))
 				$date .= " to $end";
-			$out[] = "<{$this->config['Date Tag']}>$date</{$this->config['Date Tag']}>";
-			$out[] = "<{$this->config['Title Tag']}>{$row['title']}</{$this->config['Title Tag']}>";
-			if ($this->config['Events have Location'] == 'yes')
-				$out[] = "<{$this->config['Location Tag']}>{$row['location']}</{$this->config['Location Tag']}>";
-			$out[] = "<{$this->config['Description Tag']}>{$row['description']}</{$this->config['Description Tag']}>";
+			$out[] = "<{$this->GetConfigValue('Date Tag')}>$date</{$this->GetConfigValue('Date Tag')}>";
+			$out[] = "<{$this->GetConfigValue('Title Tag')}>{$row['title']}</{$this->GetConfigValue('Title Tag')}>";
+			if ($this->GetConfigValue('Events have Location') == 'yes')
+				$out[] = "<{$this->GetConfigValue('Location Tag')}>{$row['location']}</{$this->GetConfigValue('Location Tag')}>";
+			$out[] = "<{$this->GetConfigValue('Description Tag')}>{$row['description']}</{$this->GetConfigValue('Description Tag')}>";
 			$cal[] = join("\r\n",$out);
 		}
-		return "<div class=\"Calendar\">".join($this->config['Event Separator'],$cal)."</div>";
+		return "<div class=\"Calendar\">".join($this->GetConfigValue('Event Separator'),$cal)."</div>";
 	}
 	
 	function AdminMenuText()
@@ -180,26 +188,24 @@ class Calendar extends PluginBase
 	function RenderAdmin()
 	{
 		$ds = new DataSet('calendar','id');
-		$ds->AddColumns('id','plugin','instance','start','end','title','location','description');
-		$ds->AddDataFilter('plugin',$this->pid);
+		$ds->AddColumns('id','instance','start','end','title','location','description');
 		$ds->AddDataFilter('instance',$this->iid);
 		$dg = new DataGrid($ds);
 		$dg->AddColumn("Start Date",'start');
-		if ($this->config['Allow Multi-day Events'] == 'yes')
+		if ($this->GetConfigValue('Allow Multi-day Events') == 'yes')
 			$dg->AddColumn('End Date','end');
 		$dg->AddColumn('Title','title');
 		$dg->AddEditColumn();
 		$dg->AddDeleteColumn();
 		$dg->insertlabel='Add a new Event';
 		$dg->AddUnixDateEditor('start','Start Date:');
-		if ($this->config['Allow Multi-day Events'] == 'yes')
+		if ($this->GetConfigValue('Allow Multi-day Events') == 'yes')
 			$dg->AddUnixDateEditor('end','End Date:');
 		else 
 			$ds->OnBeforeUpdate = $ds->OnBeforeInsert = "CalendarSetEndDate";
-		$dg->AddConstant('plugin',$this->pid);
 		$dg->AddConstant('instance',$this->iid);
 		$dg->AddInput('title','Title:',60,60);
-		if ($this->config['Events have Location'] == 'yes')
+		if ($this->GetConfigValue('Events have Location') == 'yes')
 		{
 			$dg->AddColumn('','location','');
 			$dg->AddInput('location','Location:',60,60);
